# openapi.yml
openapi: 3.0.3
info:
  title: Meye Core API
  description: |
    **Meye Core API** is a comprehensive role-playing game (RPG) campaign management system built with Clean Architecture principles.

    ## Key Features
    - **User Management**: Support for multiple user roles (Admin, Master, Player)
    - **Campaign Management**: Create and manage RPG campaigns with game masters
    - **Player Characters (PJs)**: Comprehensive character creation and stat management
    - **Session Tracking**: Record game sessions with XP assignments
    - **Event-Driven**: Domain events published to RabbitMQ for async processing

    ## Authentication
    Most endpoints require JWT Bearer token authentication. Obtain a token via the `/api/v1/users/login` endpoint.

    ## Authorization
    The API implements role-based access control (RBAC):
    - **Admin**: Can create new users
    - **Master**: Can create campaigns, invite users, and manage sessions
    - **Player**: Can create and manage their own player characters

    ## Architecture
    This API follows Clean Architecture patterns with:
    - Domain-driven design (DDD)
    - CQRS patterns for use cases
    - Event sourcing for domain events
    - PostgreSQL for persistence
    - RabbitMQ for event publishing

  version: 1.0.0
  contact:
    name: Meye Core API Support
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://api.meye.example.com
...
paths:
...
  /api/v1/users:
    post:
      tags:
        - Users
      summary: Create new user
      description: |
        Create a new user in the system. Only accessible by users with the **Admin** role.

        **User Roles:**
        - `admin`: System administrator with full access
        - `master`: Campaign master who can create and manage campaigns
        - `player`: Regular player who can create and manage their own characters

        **Password Requirements:**
        - Minimum 8 characters
        - Maximum 72 characters
        - Will be hashed using bcrypt before storage
      operationId: createUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Insufficient permissions (requires Admin role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              examples:
                conflict:
                  value:
                    error: Username already exists
                    code: USERNAME_EXISTS
  /api/v1/users/self:
    get:
      tags:
        - Users
      summary: Gets user details
      description: |
        Gets the details of the logged user.
      operationId: getUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Details of the logged user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from the login endpoint.
        Include it in the Authorization header:
        ```
        Authorization: Bearer <your_jwt_token>
        ```
...
  schemas:
...
    CreateUserRequest:
      type: object
      required:
        - username
        - password
        - role
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 100
          pattern: '^[a-zA-Z0-9]+$'
          description: Alphanumeric username (no special characters)
          example: masteruser
        password:
          type: string
          minLength: 8
          maxLength: 72
          format: password
          description: User password (will be hashed with bcrypt)
          example: SecurePassword123
        role:
          $ref: '#/components/schemas/UserRole'
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier
          example: 123e4567-e89b-12d3-a456-426614174000
        username:
          type: string
          description: Unique username
          example: masteruser
        role:
          $ref: '#/components/schemas/UserRole'
    UserRole:
      type: string
      enum:
        - admin
        - master
        - player
      description: |
        User role defining permissions:
        - **admin**: System administrator, can create users
        - **master**: Campaign master, can create campaigns and sessions
        - **player**: Regular player, can create and manage characters
      example: master
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
          example: Invalid request parameters
        code:
          type: string
          description: Machine-readable error code
          example: VALIDATION_ERROR
  responses:
    BadRequest:
      description: Invalid request parameters or validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            validationError:
              value:
                error: 'Key: ''CreateUserInputBody.Password'' Error:Field validation for ''Password'' failed on the ''min'' tag'
                code: VALIDATION_ERROR
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            missingToken:
              value:
                error: Authentication required
                code: UNAUTHORIZED
            invalidToken:
              value:
                error: Invalid or expired token
                code: INVALID_TOKEN
...